# JWT validation perf - string[] to HashSet<string>