--- a/libraries/Microsoft.Bot.Connector/Authentication/AuthenticationConstants.cs
+++ b/libraries/Microsoft.Bot.Connector/Authentication/AuthenticationConstants.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Collections.Generic;
+
 namespace Microsoft.Bot.Connector.Authentication
 {
     /// <summary>
@@ -42,7 +44,7 @@ namespace Microsoft.Bot.Connector.Authentication
         /// Allowed token signing algorithms. Tokens come from channels to the bot. The code
         /// that uses this also supports tokens coming from the emulator.
         /// </summary>
-        public static readonly string[] AllowedSigningAlgorithms = new[] { "RS256", "RS384", "RS512" };
+        public static readonly HashSet<string> AllowedSigningAlgorithms = new HashSet<string>(new [] { "RS256", "RS384", "RS512" });
 
         /// <summary>
         /// "azp" Claim.