There were a couple scenarios in JWT validation using `string[]`, thus
O(N) cost, vs using `HashSet<string>` (for O(1) cost.


``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3110.0
  Job-KGFYSY : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3110.0

Jit=RyuJit  Platform=X86  Runtime=Clr  

```
|                Method |     Mean |     Error |    StdDev | Scaled | ScaledSD | Allocated |
|---------------------- |---------:|----------:|----------:|-------:|---------:|----------:|
|   LookupInArray_Found | 70.50 ns | 1.5900 ns | 4.6632 ns |   1.00 |     0.00 |       0 B |
| LookupInHashSet_Found | 28.27 ns | 0.5958 ns | 0.8918 ns |   0.40 |     0.03 |       0 B |


``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3110.0
  Job-KGFYSY : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3110.0

Jit=RyuJit  Platform=X86  Runtime=Clr  

```
|                   Method |     Mean |     Error |    StdDev | Scaled | Allocated |
|------------------------- |---------:|----------:|----------:|-------:|----------:|
|   LookupInArray_NotFound | 73.61 ns | 1.5145 ns | 2.2200 ns |   1.00 |       0 B |
| LookupInHashSet_NotFound | 18.62 ns | 0.4312 ns | 0.9374 ns |   0.25 |       0 B |